requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
cycler==0.11.0
defusedxml==0.7.1
Django==3.2.9
django-allauth==0.47.0
django-crispy-forms==1.13.0
django-markdownx==3.0.1
idna==3.3
importlib-metadata==4.9.0
kiwisolver==1.3.2
Markdown==3.3.6
matplotlib==3.4.3
numpy==1.21.3
oauthlib==3.1.1
Pillow==8.4.0
psycopg2-binary==2.9.2
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.4
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
zipp==3.6.0
